% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormGen.R
\name{NormGen}
\alias{NormGen}
\alias{NormGen-package}
\title{The list of package functions and their demonstration}
\description{
The \strong{NormGen} presents the normal pseudo-random number.
The list of package functions is as follows:
}
\section{Function for the normal distribution}{


\code{\link{Rnorm}}
}
