% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallel_Rnorm}
\alias{parallel_Rnorm}
\title{Parallel Pseudo-Normal Random Number Generator}
\usage{
parallel_Rnorm(n, mean = 0, sd = 1, seed = 12345L)
}
\arguments{
\item{n}{number of samples}

\item{mean}{distribution mean}

\item{sd}{distribution standard deviation}

\item{seed}{generator seed}
}
\description{
Generates pseudo-random numbers from a normal distribution using multiple threads.
}
