% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/Rnorm.R
\name{Rnorm}
\alias{Rnorm}
\title{Single-threaded Pseudo-Normal Random Number Generator}
\usage{
Rnorm(n, m = 0, s = 1)
}
\arguments{
\item{n}{number of observations.}

\item{m}{vector of means.}

\item{s}{vector of standard deviations.}
}
\value{
The function returns the value of random generation for the normal distribution with m equal to mean and standard deviation equal to s.
}
\description{
Original single-threaded implementation of the NormGen quantile mapping algorithm.

Random generation for the normal distribution with parameters m and s.
}
\examples{
Rnorm(10,2,2)
Rnorm(33)

}
\references{
{Sulewski P., Drapella A. (2026). \emph{An Efficient and Precise Quantile Mapping for Normal Pseudo-Random Number Generation.} Journal of Statistical Software, Under review.}
}
